Fixed clang warning compiling C-exts.

+++ lib/19/syck/ext/syck.c
@@ -518,7 +518,7 @@ syck_default_error_handler( SyckParser *p, const char *msg )
 {
     printf( "Error at [Line %d, Col %d]: %s\n",
         p->linect,
-        p->cursor - p->lineptr,
+        (int)(p->cursor - p->lineptr),
         msg );
 }
